Refactor app/pages indexers & add auto-detect of appDir
